requirements.txt
python-dotenv
colorama
beautifulsoup4
